Move some of the Next.js example files into a src/lib subtree
