+++ examples/nextjs/app/auth/signin/anonymous/route.ts
@@ -1,5 +1,5 @@
 import { anonymous } from "@convex-dev/auth/providers/anonymous/server";
-import { auth } from "@/serverAuth";
+import { auth } from "@/src/lib/serverAuth";
 import { api } from "@/convex/_generated/api";
 
 // Runs anonymous sign-in on the server: mints the session, moves the refresh
