+++ examples/nextjs/src/lib/convexAuth.tsx
