+++ examples/nextjs/src/lib/serverAuth.ts
